FIX move everythin into single file
