--- submissions/image_processing/iou.py
@@ -1,81 +1,0 @@
-from __future__ import division
-
-import math
-
-__all__ = ["cc_iou"]
-
-
-def cc_iou(circle1, circle2):
-    """
-    Intersection over Union (IoU) between two circles
-
-    Parameters
-    ----------
-    circle1 : tuple of floats
-        first circle parameters (x_pos, y_pos, radius)
-    circle2 : tuple of floats
-        second circle parameters (x_pos, y_pos, radius)
-
-    Returns
-    -------
-    float
-        ratio between area of intersection and area of union
-
-    """
-    x1, y1, r1 = circle1
-    x2, y2, r2 = circle2
-
-    d = math.hypot(x2 - x1, y2 - y1)
-
-    area_intersection = cc_intersection(d, r1, r2)
-    area_union = math.pi * (r1 * r1 + r2 * r2) - area_intersection
-
-    return area_intersection / area_union
-
-
-def cc_intersection(dist, rad1, rad2):
-    """
-    Area of intersection between two circles
-
-    Parameters
-    ----------
-    dist : positive float
-        distance between circle centers
-    rad1 : positive float
-        radius of first circle
-    rad2 : positive float
-        radius of second circle
-
-    Returns
-    -------
-    intersection_area : positive float
-        area of intersection between circles
-
-    References
-    ----------
-    http://mathworld.wolfram.com/Circle-CircleIntersection.html
-
-    """
-    if dist < 0:
-        raise ValueError("Distance between circles must be positive")
-    if rad1 < 0 or rad2 < 0:
-        raise ValueError("Circle radius must be positive")
-
-    if dist == 0 or (dist <= abs(rad2 - rad1)):
-        return min(rad1, rad2) ** 2 * math.pi
-
-    if dist > rad1 + rad2 or rad1 == 0 or rad2 == 0:
-        return 0
-
-    rad1_sq = rad1 * rad1
-    rad2_sq = rad2 * rad2
-
-    circle1 = rad1_sq * math.acos((dist * dist + rad1_sq - rad2_sq) /
-                                  (2 * dist * rad1))
-    circle2 = rad2_sq * math.acos((dist * dist + rad2_sq - rad1_sq) /
-                                  (2 * dist * rad2))
-    intersec = 0.5 * math.sqrt((-dist + rad1 + rad2) * (dist + rad1 - rad2) *
-                               (dist - rad1 + rad2) * (dist + rad1 + rad2))
-    intersection_area = circle1 + circle2 + intersec
-
-    return intersection_area
